third_party/WebKit/LayoutTests/svg/dynamic-updates/SVGGElement-svgdom-requiredFeatures-expected.txt

SVG 1.1 dynamic update tests

Check that SVGGElement is initially displayed
PASS gElement.firstElementChild.getBBox().width is 200
PASS gElement.firstElementChild.getBBox().height is 200
Check that setting requiredFeatures to something invalid makes it not render
PASS gElement.firstElementChild.getBBox().width is 0
PASS gElement.firstElementChild.getBBox().height is 0
Check that setting requiredFeatures to something valid makes it render again
PASS gElement.firstElementChild.getBBox().width is 200
PASS gElement.firstElementChild.getBBox().height is 200
Check that adding something valid to requiredFeatures keeps rendering the element
PASS gElement.firstElementChild.getBBox().width is 200
PASS gElement.firstElementChild.getBBox().height is 200
Check that adding something invalid to requiredFeatures makes it not render
PASS gElement.firstElementChild.getBBox().width is 0
PASS gElement.firstElementChild.getBBox().height is 0
PASS successfullyParsed is true

TEST COMPLETE
